Rework bulk-action dropdown filter as hook_form_alter instead of a Views field-plugin substitution
The Views field-plugin approach (MukurtuAccessFilteredBulkForm, wired in
via a plugin_id config swap) passed its Kernel test but silently never
executed during real Views rendering -- confirmed live on a real site: the
form always instantiated VBO's own unfiltered field class regardless of
the configured plugin_id, for reasons that resisted fairly deep tracing
(config and plugin discovery were provably correct in every context
checked; the actual runtime field handler still wasn't the swapped class).

Replaced with a hook_form_alter() that strips inaccessible options
directly from the built form array, matching a proven, standard mechanism
already used elsewhere in this codebase for the same view's other bulk
actions. Verified live via HTTP with two different privilege levels
(empty dropdown for a user who can't act on anything visible; a
correctly narrowed set for a user who can). The new Kernel test drives
the *real* Views form-building path (DisplayPluginBase::render() ->
ViewsForm::create() -> FormBuilder::getForm()) rather than a
hand-constructed stub, specifically to catch this class of gap.
